chore: added toDomain() methods on jpa entities

+++ week02_arch/src/main/java/com/example/hhplus_arch_jvm/application/domain/CourseRegistrationCount.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record CourseRegistrationCount(
         Long courseId,
-        Long count,
-        Long max
+        Integer count,
+        Integer max
 ) {
 }
